implemented new fuctionality
Call StartRegisteringAllThreadAllocationsAsExpectedLeaks to register all
following allocations in the current thread as expected memory leaks.

Call StopRegisteringAllThreadAllocationsAsExpectedLeaks to stop that
behaviour.

Only one thread at a time can be in that state; if a second thread tries
to enter it, it will be blocked until the first thread exits from that
state.

Available only in EnableMemoryLeakReporting mode.